Adding propper check to see if user owns a file, and allowing them to delete if so.

+++ framework/modules/file/controllers/fileController.php
@@ -21,9 +21,9 @@ class fileController extends expController {
     public $basemodel_name = "expFile";
     //public $useractions = array('showall'=>'Show all');
     //public $add_permissions = array('picker'=>'Manage Files');
-    //public $remove_permissions = array('edit');
+    public $remove_permissions = array('delete');
     public $requires_login = array('picker'=>'must be logged in','edit_alt'=>'must be logged in');
-	public $codequality = 'beta';
+	public $codequality = 'stable';
 
     function displayname() { return "File Manager"; }
     function description() { return "Add and manage Exponent Files"; }
@@ -194,13 +194,17 @@ class fileController extends expController {
     } 
     
     public function delete() {
-        global $db;
+        global $db,$user;
         $file = new expFile($this->params['id']);
-        $file->delete();
-        if (unlink($file->directory.$file->filename)) {
-            flash('message',$file->filename.' '.gt('was successfully deleted'));
+        if ($user->id==$file->poster || $user->isAdmin()) {
+            $file->delete();
+            if (unlink($file->directory.$file->filename)) {
+                flash('message',$file->filename.' '.gt('was successfully deleted'));
+            } else {
+                flash('error',$file->filename.' '.gt('was deleted from the database, but could not be removed from the file system.'));
+            }
         } else {
-            flash('error',$file->filename.' '.gt('was deleted from the database, but could not be removed from the file system.'));
+            flash('error',$file->filename.' '.gt('wasn\'t deleted because you don\'t own the file.'));
         }
         redirect_to(array("controller"=>'file',"action"=>'picker',"ajax_action"=>1,"update"=>$this->params['update'],"fck"=>$this->params['fck']));
     }
